Reduce size of images.
The base size of the virtual disk is reduced to 4 GB; if you need more disk
space, you can add a second disk to the VM in Vagrant, and add that to the
pool. (Exmaple forthcoming.)

Also, use the AHCI disk controller interface for a huge bonus in disk speed.

I also tried activating the SSD feature in VirtualBox, but that made the VM
becode unstable and lock up, including SATA errors inside the VM.